Legger til bostedsadresse på person og gjør den optional
