+++ src/main/java/no/nav/familie/ks/sak/app/rest/behandling/grunnlag/personinformasjon/RestPerson.kt
@@ -1,5 +1,6 @@
 package no.nav.familie.ks.sak.app.rest.behandling.grunnlag.personinformasjon
 
+import no.nav.familie.ks.sak.app.behandling.domene.grunnlag.personopplysning.Bostedsadresse
 import no.nav.familie.ks.sak.app.behandling.domene.grunnlag.personopplysning.Person
 
 import java.time.LocalDate
@@ -9,6 +10,7 @@ data class RestPerson internal constructor(
         val navn: String,
         val kjønn: String?,
         val fødselsdato: LocalDate,
+        val bostedsadresse: Bostedsadresse?,
         val statsborgerskap: String,
         val relasjoner: List<RestPersonrelasjon>,
         val personhistorikk: RestPersonhistorikk)
@@ -19,6 +21,10 @@ fun Person.toRestPerson() = RestPerson(
         navn = this.navn,
         kjønn = this.kjønn,
         fødselsdato = this.fødselsdato,
+        bostedsadresse = when(this.bostedsadresse.isPresent){
+            true -> this.bostedsadresse.get()
+            false -> null
+        },
         statsborgerskap = this.statsborgerskap.kode,
         relasjoner = this.relasjoner.map { it.toRestPersonrelasjon() },
         personhistorikk = this.toRestPersonhistorikk()
